fix unnecessary type casting

+++ src/optimum/rbln/ops/moe.py
@@ -23,12 +23,11 @@ def compute_masked_routing_weight_softmax_first(router_logits: Tensor, top_k: in
     #   renormalize=False: softmax → topk → scatter           (pre_norm)
     router_logits_t = router_logits.transpose(0, 1)  # [T, E] -> [E, T]
     if renormalize:
-        topk_values, topk_ids = torch.topk(router_logits_t.to(torch.float32), top_k, dim=0)
-        topk_weights = torch.softmax(topk_values, dim=0).to(router_logits.dtype)
+        topk_values, topk_ids = torch.topk(router_logits_t, top_k, dim=0)
+        topk_weights = torch.softmax(topk_values, dim=0)
     else:
-        routing = torch.softmax(router_logits_t.to(torch.float32), dim=0)
+        routing = torch.softmax(router_logits_t, dim=0)
         topk_weights, topk_ids = torch.topk(routing, top_k, dim=0)
-        topk_weights = topk_weights.to(router_logits.dtype)
     masked = torch.zeros_like(router_logits_t, dtype=router_logits.dtype)
     masked.scatter_(0, topk_ids, topk_weights)
     return masked  # [E, T]
@@ -39,7 +38,7 @@ def compute_masked_routing_weight_topk_first(router_logits: Tensor, top_k: int) 
     # Same [E, T] / dim=0 layout as softmax_first for compiler pattern matching.
     router_logits_t = router_logits.transpose(0, 1)  # [T, E] -> [E, T]
     topk_values, topk_ids = torch.topk(router_logits_t, top_k, dim=0)
-    topk_weights = torch.softmax(topk_values.to(torch.float32), dim=0).to(router_logits.dtype)
+    topk_weights = torch.softmax(topk_values, dim=0)
     masked = torch.zeros_like(router_logits_t, dtype=router_logits.dtype)
     masked.scatter_(0, topk_ids, topk_weights)
     return masked  # [E, T]
